# iCanteen-api接口.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "iCanteen-api接口",
    "description": "校园“智汇”食堂后端接口文档（返回统一使用 Result 包装）",
    "version": "1.0.0"
  },
...
  "paths": {
    "/user/code": {
      "post": {
        "summary": "发送登录验证码",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [
          {
            "name": "phone",
            "in": "query",
            "description": "手机号（11位大陆手机号）",
            "required": true,
            "example": "13456789004",
            "schema": {
              "type": "string",
              "pattern": "^1[3-9]\\d{9}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/login": {
      "post": {
        "summary": "登录（验证码或密码二选一）",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginFormDTO"
              },
              "examples": {
                "短信登录": {
                  "value": {
                    "phone": "13456789004",
                    "code": "545547"
                  },
                  "summary": "短信登录"
                },
                "密码登录": {
                  "value": {
                    "phone": "13800000002",
                    "password": "123456"
                  },
                  "summary": "密码登录"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result(data=token)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/me": {
      "get": {
        "summary": "获取当前登录用户",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "aa19515bcc324ccb811ac30548d0f520",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=UserDTO)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/info": {
      "put": {
        "summary": "修改当前登录用户详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "aa19515bcc324ccb811ac30548d0f520",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInfoUpdateDTO"
              },
              "example": {
                "birthday": "2003-12-31",
                "city": "Shenzhen",
                "gender": false,
                "phone": "13456789048",
                "icon": "/imgs/icons/default2.png",
                "introduce": "hihihi!",
                "nickname": "user_abc",
                "password": "321cba"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "获取当前登录用户详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "aa19515bcc324ccb811ac30548d0f520",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=UserProfileDTO)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/logout": {
      "post": {
        "summary": "退出登录",
        "deprecated": false,
        "description": "",
        "tags": [
          "用户"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "登录令牌，支持 token 或 Bearer token",
            "required": false,
            "example": "00be94d772df4bd1b84fc2a4a1b3f918",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/sign": {
      "post": {
        "summary": "用户签到",
        "deprecated": false,
        "description": "登录用户执行当天签到（Bitmap），首次签到成功奖励1积分。",
        "tags": [
          "用户"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "aa19515bcc324ccb811ac30548d0f520",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/user/sign/count": {
      "get": {
        "summary": "查询当前连续签到天数",
        "deprecated": false,
        "description": "返回当前登录用户截至今天的连续签到天数",
        "tags": [
          "用户"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "a59e4a4744b94887a7126fc350cd640d",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=连续签到天数)",
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/food-order": {
      "post": {
        "summary": "用户下单",
        "deprecated": false,
        "description": "登录用户创建餐品订单，后端进行并发库存扣减与事务一致性控制。",
        "tags": [
          "订单"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "9653930b1b53492a80214bbca92c77a3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodOrderCreateDTO"
              },
              "example": {
                "items": [
                  {
                    "dishId": 7,
                    "quantity": 1
                  },
                  {
                    "dishId": 8,
                    "quantity": 2
                  }
                ],
                "remark": "不要葱和香菜，谢谢！",
                "longitude": 120.151506,
                "latitude": 30.333422
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Result(data=orderId)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/food-order/my": {
      "get": {
        "summary": "用户分页查询我的订单",
        "deprecated": false,
        "description": "",
        "tags": [
          "订单"
        ],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "0待接单,1制作中,2待取餐,3已完成,4已取消",
            "required": false,
            "example": 0,
            "schema": {
              "type": "integer",
              "enum": [
                0,
                1,
                2,
                3,
                4
              ]
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "9653930b1b53492a80214bbca92c77a3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=FoodOrderSimpleDTO[],total=总数)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/food-order/{id}": {
      "get": {
        "summary": "用户查询订单详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "订单"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 2595113729524213,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "9653930b1b53492a80214bbca92c77a3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=FoodOrderDetailDTO)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/food-order/{id}/cancel": {
      "put": {
        "summary": "用户取消订单",
        "deprecated": false,
        "description": "",
        "tags": [
          "订单"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 2595113729524213,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "cancelReason",
            "in": "query",
            "description": "",
            "required": false,
            "example": "点错了呜呜呜",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "用户登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "9653930b1b53492a80214bbca92c77a3",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/admin/food-order/{id}": {
      "get": {
        "summary": "管理员查询订单详情",
        "deprecated": false,
        "description": "",
        "tags": [
          "订单-管理"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 2342921839837993,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "管理员登录令牌，支持 token 或 Bearer token",
            "required": true,
            "example": "0406b05b0f534d0ca648b305972d0fa4",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result(data=FoodOrderDetailDTO)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "Result": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "errorMsg": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "null"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        }
      },
      "LoginFormDTO": {
        "type": "object",
        "description": "登录参数：phone 必填，code 与 password 二选一且只能传一个",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^1[3-9]\\d{9}$",
            "description": "手机号（必填）"
          },
          "code": {
            "type": "string",
            "description": "短信验证码（与 password 二选一）"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "登录密码（与 code 二选一）"
          }
        },
        "required": [
          "phone"
        ]
      },
...
      "UserInfoUpdateDTO": {
        "type": "object",
        "description": "修改本人详情，允许修改 phone、password、nickname、icon、city、introduce、gender、birthday",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^1[3-9]\\d{9}$"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "nickname": {
            "type": "string",
            "maxLength": 32
          },
          "icon": {
            "type": "string",
            "maxLength": 255
          },
          "city": {
            "type": "string",
            "maxLength": 64
          },
          "introduce": {
            "type": "string",
            "maxLength": 256
          },
          "gender": {
            "type": "boolean",
            "nullable": true
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "nullable": true
          }
        }
      },
...
      "FoodOrderCreateDTO": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/FoodOrderItemCreateDTO"
            }
          },
          "remark": {
            "type": "string"
          }
        }
      },
...
      "FoodOrderItemCreateDTO": {
        "type": "object",
        "required": [
          "dishId",
          "quantity"
        ],
        "properties": {
          "dishId": {
            "type": "integer",
            "format": "int64"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
...
    },
...
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}